fix config issue
